Moved router specific logic out of sharable UI components
